third_party/blink/web_tests/external/wpt/workers/Worker-terminate-forever-during-evaluation-expected.txt

This is a testharness.js-based test.
PASS Worker is terminated during top-level script evaluation
FAIL Worker is terminated during top-level script evaluation (module) assert_unreached: onerror Reached unreachable code
PASS Worker is terminated during importScripts() call
PASS Worker is terminated during nested importScripts() call
PASS Worker is terminated during dynamic import()
PASS Worker is terminated during dynamic import() (module)
Harness: the test ran to completion.
